GSYE-295: Move AvailableMarketTypes into gsy_e_core package in order to curcumvent circular imports when implementing forward markets; add forward market types to AvailableMarketTypes
